requirements.txt
aiofiles
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
pywebpush
py-bcrypt
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
cryptography>=41.0.4,<42.0.0
dnspython==2.6.1
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.1.2
fastapi==0.111.0
fastapi-admin==1.0.4
fastapi-cli==0.0.4
fastapi-users==13.0.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
Jinja2==3.1.4
makefun==1.15.3
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.6
passlib==1.7.4
pendulum==3.0.0
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
time-machine==2.14.2
tortoise-orm==0.21.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
loguru
aiofiles
zxing
redis